requirements.txt
Flask==2.3.3
gunicorn==21.2.0
Pillow==10.0.0
tensorflow-cpu==2.13.0